+++ packages/core/lib/json-payload.js
@@ -15,13 +15,11 @@ Serialized payload was ${payload.length / 10e5}MB (limit = 1MB)
 metadata was removed`
     }
     payload = jsonStringify(event, null, null, { redactedPaths: EVENT_REDACTION_PATHS, redactedKeys })
-    if (payload.length > 10e5) throw new Error('payload exceeded 1MB limit')
   }
   return payload
 }
 
-module.exports.session = (event, redactedKeys) => {
-  const payload = jsonStringify(event, null, null)
-  if (payload.length > 10e5) throw new Error('payload exceeded 1MB limit')
+module.exports.session = (session, redactedKeys) => {
+  const payload = jsonStringify(session, null, null)
   return payload
 }